-Generate pages based on version for the history views and downloads for go-mono.com (mk-index and mk-sources)   ./upload-prod now requires a version   Don't generate repo info or zip files for archive versions
-Consolidate the stylesheet into packaging.css
-Generate links to spec files if they exist (which will in the future)
-Fix bug so pages for USE_ZIP_PKG don't get generated (mk-index)


svn path=/trunk/release/; revision=53324

+++ website/mk-index.sh
@@ -5,30 +5,38 @@ packagingdir=$confdir/../packaging
 . $packagingdir/shared-code.sh
 
 
-
 WEB_DIR=$1
+VERSION=$2
+
+if test x$VERSION = x; then
+	echo "Usage: mk-index.sh <web_dir> <version>"
+	exit 1
+fi
 
 cd $WEB_DIR
 
 for distro_conf in $packagingdir/conf/*-*-*; do
 
 	# Skip the distros that use zip packaging system
-	! egrep "^USE_ZIP_PKG" $packagingdir/conf/$distro_conf > /dev/null 2>&1 || continue
+	! egrep "^USE_ZIP_PKG" $distro_conf > /dev/null 2>&1 || continue
 	
 	distro_info `basename $distro_conf`
 	
 	mkdir -p $DISTRO
+	mkdir -p $WEB_DIR/../archive/$VERSION/download/$DISTRO
 	
 	OUT=$DISTRO/index.html
-	
+	HISTORY_OUT=$WEB_DIR/../archive/$VERSION/download/$DISTRO/index.html
 
 	rm -rf $OUT
+	rm -rf $HISTORY_OUT
 	
 	cat $confdir/groups | while read line
 	do
 		if [ "x${line:0:1}" == "x#" ]; then
 			ARGS=(${line:1})
 			RPMS=()
+			SPECS=()
 			for package in ${ARGS[@]:1}; do 
 				. $packagingdir/defs/$package
 				
@@ -41,10 +49,18 @@ for distro_conf in $packagingdir/conf/*-*-*; do
 					[[ $i == *.src.rpm ]] && continue
 					RPMS=(${RPMS[@]} $i)
 				done
+
+				for i in $LATEST_VERSION/*.spec; do
+					if [ -e $i ] ; then
+						SPECS=(${SPECS[@]} $i)
+					fi
+				done
+
 			done
 			
 			if [ ${#RPMS[@]} -eq 0 ]; then
 				echo "<p>Not available for this platform</p>" >> $OUT
+				echo "<p>Not available for this platform</p>" >> $HISTORY_OUT
 				continue
 			fi
 			
@@ -54,24 +70,48 @@ for distro_conf in $packagingdir/conf/*-*-*; do
 			# rpms are compressed anyways -- doing any compression is a waste of time
 			zip -j -0 $DISTRO/$zipname.zip ${RPMS[@]}
 			
-			echo "<p><a href='$zipname.zip'><img src='/zip-icon.png' />All of these files in a ZIP file</a></p>" >> $OUT
+			echo "<p><a href='$zipname.zip'><img src='/zip-icon.png' />All of these RPMs in a ZIP file</a></p>" >> $OUT
 			echo "<ul>" >> $OUT
+			echo "<ul>" >> $HISTORY_OUT
 			
 			for i in ${RPMS[@]}; do
 				NAME=$(rpm_query NAME $i)
 				DESC=$(rpm_query SUMMARY $i)
 				echo "<li><a href='../$i'>$NAME</a> -- $DESC</li>" >> $OUT
+				echo "<li><a href='../../../../download/$i'>$NAME</a> -- $DESC</li>" >> $HISTORY_OUT
 			done
 			
 			echo "</ul>" >> $OUT
+			echo "</ul>" >> $HISTORY_OUT
+
+			# Print links to spec files
+			if [ ${#SPECS[@]} -eq 0 ]; then
+				continue
+			fi
+
+			echo "<p>RPM Spec files: " >> $OUT
+			echo "<p>RPM Spec files: " >> $HISTORY_OUT
+			for i in ${SPECS[@]}; do
+				NAME=$(basename $i)
+				echo "<a href='../$i'>$NAME</a> " >> $OUT
+				echo "<a href='../../../../download/$i'>$NAME</a> " >> $HISTORY_OUT
+			done
+			echo "</p>" >> $OUT
+			echo "</p>" >> $HISTORY_OUT
+
 			
 		elif [ "x${line:0:1}" == "x!" ]; then
 			line=$(echo ${line:1})
 			
 			. $confdir/$line >> $OUT
-			
+
+			# Don't display repository information for history pages
+			if test x$line != xget_rpm_install ; then
+				. $confdir/$line >> $HISTORY_OUT
+			fi
 		else
 			echo ${line//\\[\\[arch\\]\\]/$(basename $distro_conf)} >> $OUT
+			echo ${line//\\[\\[arch\\]\\]/$(basename $distro_conf)} >> $HISTORY_OUT
 		fi
 	done
 done
